requirements.txt
streamlit==1.31.1
langchain==0.1.8
langchain-community==0.0.24
langchain-core==0.1.26
pyperclip-1.8.2
